pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "japim"
version = "0.1.0"
description = "PaddleOCR based PDF image PII masking pipeline with REST API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116,<1.0",
  "uvicorn[standard]>=0.35,<1.0",
  "python-multipart>=0.0.20,<1.0",
  "PyYAML>=6.0,<7.0",
  "pydantic>=2.7,<3.0",
  "numpy>=1.26,<3.0",
  "Pillow>=10.0,<12.0",
  "PyMuPDF>=1.24,<2.0",
  "opencv-python-headless>=4.10,<5.0",
  "pandas>=2.2,<3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "httpx>=0.28,<1.0",
]
ocr = [
  "paddleocr>=3.4.0,<4.0",
]

[project.scripts]
japim = "japim.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
